When run training in Cluster Mode Spark with Apache Zeppelin, suspect this one killed SparkContext when training failed.
I try to run XGBoost Training with Data Frame and somehow when training failed, SparkContext also been closed which usually is not 
a problem when SparkContext is not shared with others but when under Apache Zeppelin and shared mode, it will kill shared context which is not necessary. I'm not 100% sure if the code I comment out is the root cause or not so please correct me if that is the wrong place. Can't find a clear call for close form the JVM package so the only thing I think about is exception been throw out here causing the issue.